requirements.txt
alembic==1.8.1
aniso8601==9.0.1
asgiref==3.5.2
atomicwrites==1.4.0
attrs==21.2.0
beautifulsoup4==4.10.0
black==21.12b0
bs4==0.0.1
CacheControl==0.12.11
cachy==0.3.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
cleo==0.8.1
click==8.0.3
clikit==0.6.2
colorama==0.4.4
coverage==6.4.1
crashtest==0.3.1
DateTime==4.5
distlib==0.3.5
Django==4.1
django-nose==1.4.7
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
factory-boy==3.2.1
Faker==11.3.0
filelock==3.8.0
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
flask-swagger-ui==4.11.1
greenlet==1.1.2
html5lib==1.1
idna==3.3
importlib-metadata==4.8.1
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.2
jsonschema==4.1.2
keyring==23.8.2
lockfile==0.12.2
Mako==1.2.1
MarkupSafe==2.1.1
marshmallow==3.14.1
marshmallow-dataclass==8.5.8
msgpack==1.0.4
mypy==0.961
mypy-extensions==0.4.3
nose==1.3.7
packaging==20.9
parameterized==0.8.1
pastel==0.2.1
pathspec==0.9.0
pexpect==4.8.0
Pillow==9.2.0
pinocchio==0.4.3
pkginfo==1.8.3
platformdirs==2.4.1
pluggy==1.0.0
poetry==1.1.14
poetry-core==1.0.8
psycopg2==2.9.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
py==1.11.0
PyJWT==2.3.0
pylev==1.4.0
pyparsing==3.0.9
pyrsistent==0.18.0
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-factoryboy==2.5.0
python-dateutil==2.8.2
pytz==2022.2.1
pywin32-ctypes==0.2.0
requests==2.28.0
requests-toolbelt==0.9.1
shellingham==1.5.0
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.31
sqlparse==0.4.2
text-unidecode==1.3
tomli==1.2.3
tomlkit==0.11.3
typing-inspect==0.7.1
typing_extensions==4.0.1
tzdata==2022.2
urllib3==1.26.9
virtualenv==20.16.3
webencodings==0.5.1
Werkzeug==2.0.2
zipp==3.6.0
zope.interface==5.4.0
